requirements.txt
pytesseract
pyautogui
opencv-python
Pillow